# openapi.json
{"openapi":"3.1.0","info":{"title":"ML Platform Backend","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/datasets":{"get":{"summary":"List Datasets","description":"List all uploaded datasets from database","operationId":"list_datasets_datasets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/datasets/{dataset_id}":{"get":{"summary":"Get Dataset","description":"Get a specific dataset by ID from database","operationId":"get_dataset_datasets__dataset_id__get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Update Dataset","description":"Update dataset metadata","operationId":"update_dataset_datasets__dataset_id__put","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","title":"Updates"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Dataset","description":"Delete a dataset from both file system and database","operationId":"delete_dataset_datasets__dataset_id__delete","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasets/{dataset_id}/data":{"put":{"summary":"Update Dataset Data","description":"Update dataset data (columns and rows) - saves to CSV file","operationId":"update_dataset_data_datasets__dataset_id__data_put","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetDataUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasets/{dataset_id}/preview":{"get":{"summary":"Get Dataset Preview","description":"Get dataset preview with pagination","operationId":"get_dataset_preview_datasets__dataset_id__preview_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Page Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/upload":{"post":{"summary":"Upload Dataset","description":"Upload a dataset file and store metadata in database","operationId":"upload_dataset_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_dataset_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/validate/dataset":{"post":{"summary":"Validate Dataset Endpoint","description":"Run comprehensive dataset validation (legacy endpoint)","operationId":"validate_dataset_endpoint_validate_dataset_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasets/{dataset_id}/validate":{"post":{"summary":"Validate Dataset By Id","description":"Run comprehensive dataset validation by dataset ID - supports numerical, categorical, and mixed datasets","operationId":"validate_dataset_by_id_datasets__dataset_id__validate_post","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationRequestWithId"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasets/{dataset_id}/validations":{"get":{"summary":"Get Dataset Validations","description":"Get validation history for a dataset","operationId":"get_dataset_validations_datasets__dataset_id__validations_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasets/{dataset_id}/validations/{validation_id}":{"get":{"summary":"Get Validation Detail","description":"Get detailed validation result by validation ID","operationId":"get_validation_detail_datasets__dataset_id__validations__validation_id__get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}},{"name":"validation_id","in":"path","required":true,"schema":{"type":"integer","title":"Validation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasets/{dataset_id}/preprocessing":{"get":{"summary":"Get Dataset Preprocessing Steps","description":"Get preprocessing history for a dataset","operationId":"get_dataset_preprocessing_steps_datasets__dataset_id__preprocessing_get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/datasets/{dataset_id}/preprocessing/{step_id}":{"get":{"summary":"Get Preprocessing Step Detail","description":"Get detailed preprocessing step by step ID","operationId":"get_preprocessing_step_detail_datasets__dataset_id__preprocessing__step_id__get","parameters":[{"name":"dataset_id","in":"path","required":true,"schema":{"type":"string","title":"Dataset Id"}},{"name":"step_id","in":"path","required":true,"schema":{"type":"integer","title":"Step Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocess/missing-values":{"post":{"summary":"Preprocess Missing Values","description":"Handle missing values in a dataset","operationId":"preprocess_missing_values_preprocess_missing_values_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MissingValuesRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocess/data-cleaning":{"post":{"summary":"Preprocess Data Cleaning","description":"Process data cleaning operations","operationId":"preprocess_data_cleaning_preprocess_data_cleaning_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DataCleaningRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocess/categorical-encoding":{"post":{"summary":"Preprocess Categorical Encoding","description":"Process categorical encoding operations","operationId":"preprocess_categorical_encoding_preprocess_categorical_encoding_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoricalEncodingRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocess/feature-scaling":{"post":{"summary":"Preprocess Feature Scaling","description":"Process feature scaling operations","operationId":"preprocess_feature_scaling_preprocess_feature_scaling_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureScalingRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocess/feature-selection":{"post":{"summary":"Preprocess Feature Selection","description":"Process feature selection operations","operationId":"preprocess_feature_selection_preprocess_feature_selection_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureSelectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocess/feature-extraction":{"post":{"summary":"Preprocess Feature Extraction","description":"Process feature extraction operations","operationId":"preprocess_feature_extraction_preprocess_feature_extraction_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureExtractionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preprocess/dataset-splitting":{"post":{"summary":"Preprocess Dataset Splitting","description":"Process dataset splitting operations","operationId":"preprocess_dataset_splitting_preprocess_dataset_splitting_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DatasetSplittingRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/train":{"post":{"summary":"Train Model","description":"Create a new training job","operationId":"train_model_train_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrainRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}":{"get":{"summary":"Job Status","description":"Get the status of a training job","operationId":"job_status_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}/pause":{"post":{"summary":"Pause Job","description":"Pause a running training job","operationId":"pause_job_jobs__job_id__pause_post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}/resume":{"post":{"summary":"Resume Job","description":"Resume a paused training job","operationId":"resume_job_jobs__job_id__resume_post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}/stop":{"post":{"summary":"Stop Job","description":"Stop/Cancel a training job","operationId":"stop_job_jobs__job_id__stop_post","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/model-selection":{"post":{"summary":"Model Selection","description":"Process model selection and training","operationId":"model_selection_model_selection_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModelSelectionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_upload_dataset_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_dataset_upload_post"},"CategoricalEncodingRequest":{"properties":{"dataset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"method":{"type":"string","title":"Method"},"columns":{"items":{"type":"string"},"type":"array","title":"Columns"},"target_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Column"},"drop_first":{"type":"boolean","title":"Drop First","default":false},"handle_unknown":{"type":"string","title":"Handle Unknown","default":"ignore"},"ordinal_mapping":{"anyOf":[{"additionalProperties":{"additionalProperties":{"type":"integer"},"type":"object"},"type":"object"},{"type":"null"}],"title":"Ordinal Mapping"},"n_features":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"N Features"}},"type":"object","required":["method","columns"],"title":"CategoricalEncodingRequest"},"DataCleaningRequest":{"properties":{"dataset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"method_type":{"type":"string","title":"Method Type"},"method":{"type":"string","title":"Method"},"columns":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Columns","default":[]},"strategy":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Strategy"},"constant_value":{"anyOf":[{"type":"string"},{"type":"integer"},{"type":"number"},{"type":"null"}],"title":"Constant Value"},"threshold":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Threshold"},"text_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Text Method"},"trim_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Trim Method"},"imbalance_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Imbalance Method"},"target_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Column"},"log_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Log Method"},"outlier_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Outlier Method"},"skewness_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Skewness Method"},"smooth_method":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Smooth Method"},"window_size":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Window Size"},"remove_special_chars":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Remove Special Chars","default":false},"subset":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Subset"}},"type":"object","required":["method_type","method"],"title":"DataCleaningRequest"},"DatasetDataUpdate":{"properties":{"columns":{"items":{},"type":"array","title":"Columns"},"rows":{"items":{},"type":"array","title":"Rows"},"totalRows":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Totalrows"}},"type":"object","required":["columns","rows"],"title":"DatasetDataUpdate"},"DatasetSplittingRequest":{"properties":{"dataset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"method":{"type":"string","title":"Method"},"test_size":{"type":"number","title":"Test Size","default":0.2},"validation_size":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Validation Size"},"random_state":{"type":"integer","title":"Random State","default":42},"shuffle":{"type":"boolean","title":"Shuffle","default":true},"stratify_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stratify Column"},"time_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Time Column"},"group_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Group Column"}},"type":"object","required":["method"],"title":"DatasetSplittingRequest"},"FeatureExtractionRequest":{"properties":{"dataset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"method":{"type":"string","title":"Method"},"columns":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Columns"},"n_components":{"type":"integer","title":"N Components","default":2},"target_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Column"},"variance_threshold":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Variance Threshold"},"random_state":{"type":"integer","title":"Random State","default":42},"svd_solver":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Svd Solver"},"whiten":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Whiten"},"tol":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Tol"},"solver":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Solver"},"shrinkage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Shrinkage"},"algorithm":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Algorithm"},"fun":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fun"},"max_iter":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Iter"},"n_iter":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"N Iter"},"perplexity":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Perplexity"},"early_exaggeration":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Early Exaggeration"},"learning_rate":{"anyOf":[{"type":"string"},{"type":"number"},{"type":"null"}],"title":"Learning Rate"},"n_neighbors":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"N Neighbors"},"min_dist":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Dist"},"metric":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Metric"}},"type":"object","required":["method"],"title":"FeatureExtractionRequest"},"FeatureScalingRequest":{"properties":{"dataset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"method":{"type":"string","title":"Method"},"columns":{"items":{"type":"string"},"type":"array","title":"Columns"},"feature_range":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Feature Range"},"with_mean":{"type":"boolean","title":"With Mean","default":true},"with_std":{"type":"boolean","title":"With Std","default":true},"with_centering":{"type":"boolean","title":"With Centering","default":true},"with_scaling":{"type":"boolean","title":"With Scaling","default":true},"n_quantiles":{"type":"integer","title":"N Quantiles","default":1000},"output_distribution":{"type":"string","title":"Output Distribution","default":"uniform"},"log_base":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Log Base"}},"type":"object","required":["method","columns"],"title":"FeatureScalingRequest"},"FeatureSelectionRequest":{"properties":{"dataset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"method":{"type":"string","title":"Method"},"columns":{"items":{"type":"string"},"type":"array","title":"Columns"},"target_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Column"},"n_features":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"N Features"},"threshold":{"type":"number","title":"Threshold","default":0.0},"correlation_threshold":{"type":"number","title":"Correlation Threshold","default":0.8},"alpha":{"type":"number","title":"Alpha","default":0.01}},"type":"object","required":["method","columns"],"title":"FeatureSelectionRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MissingValuesRequest":{"properties":{"dataset_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Id"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"method":{"type":"string","title":"Method"},"columns":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Columns","default":[]},"constant_value":{"anyOf":[{"type":"string"},{"type":"integer"},{"type":"number"},{"type":"null"}],"title":"Constant Value"},"threshold":{"type":"number","title":"Threshold","default":0.5}},"type":"object","required":["method"],"title":"MissingValuesRequest"},"ModelSelectionRequest":{"properties":{"dataset_path":{"type":"string","title":"Dataset Path"},"model_type":{"type":"string","title":"Model Type"},"target_column":{"type":"string","title":"Target Column"},"task_type":{"type":"string","title":"Task Type","default":"classification"},"n_neighbors":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"N Neighbors"},"n_estimators":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"N Estimators"},"max_depth":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Max Depth"},"C":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"C"},"kernel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Kernel"}},"type":"object","required":["dataset_path","model_type","target_column"],"title":"ModelSelectionRequest"},"TrainRequest":{"properties":{"dataset_path":{"type":"string","title":"Dataset Path"},"target_column":{"type":"string","title":"Target Column"},"task_type":{"type":"string","title":"Task Type","default":"classification"}},"type":"object","required":["dataset_path","target_column"],"title":"TrainRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"ValidationRequest":{"properties":{"dataset_path":{"type":"string","title":"Dataset Path"},"target_column":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Target Column"}},"type":"object","required":["dataset_path"],"title":"ValidationRequest"},"ValidationRequestWithId":{"properties":{"targetColumn":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Targetcolumn"},"dataset_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Dataset Path"},"validationLevel":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Validationlevel"}},"type":"object","title":"ValidationRequestWithId"}}}}